hyper_dmabuf: change type of ref to shared pages to unsigned long
Changed data type of reference for the group of pages to be shared
unsigned long in case it is direct representation of the memory
address.

+++ drivers/xen/hyper_dmabuf/xen/hyper_dmabuf_xen_shm.h
@@ -37,7 +37,7 @@ int xen_be_unshare_pages(void **refs_info, int nents);
 /* Maps provided top level ref id and then return array of pages containing
  * data refs.
  */
-struct page **xen_be_map_shared_pages(int lvl3_gref, int domid,
+struct page **xen_be_map_shared_pages(unsigned long lvl3_gref, int domid,
 				      int nents,
 				      void **refs_info);
